Rework example code and add READMEs to each step
The example code has been reworked to start with the `Main` actor and
then continue to supporting actors and classes. The goal here is to
make it easier to understand by introducing the high level concepts
and then moving into the details.

Each "step" now has a README that explains what the program does and
discusses the concepts introduced in the step. This should make it
easier to use the repository during workshops, because attendees can
go back and look over the concepts. It also makes it possible to
present this as a somewhat self-directed learning tool.