Clarify that local aliases must start with a lowercase letter
The text in Q3 of 01/README.md made it seem as though any variable
could start with an `_` character, but only fields can use `_`. I
changed the wording of the question to specifically say that local
variables must start with a lowercase letter and added a link to the
variables section of the tutorial. This should prevent unnecessary
confusion.

Closes #9